refactore: refactoring of auth components
